Added first-pass filtering logic:
Data structure: 
---------------------
C S J P_1 P_2 P_3
0 1 2 3   4   5

Explanation:
---------------------
0 - circuit
1 - score
2 - juggler
3 - 1st preference (circuit)
4 - 2nd preference (circuit)
5 - 3rd preference (circuit)

Strategy:
---------------------
The routine goes through list of lists in order and assigns -- in the
first iteration -- a juggler to the current circuit based on his/her
first preference.  The routine then goes through to the next list in the
ordered list of lists and a marks the jugglers that have been assigned:
P_1 is set to -1 in each case.  This way, assigned jugglers will be
filtered out of the running for the next round of placements...